climateAI-forecast-master/requirements.txt
dask==0.17.2
distributed==1.21.6
netCDF4==1.2.9
numpy==1.14.2
pandas==0.20.2
Pillow==5.1.0
scikit-learn==0.19.1
scipy==1.0.1
tqdm==4.19.5
xarray==0.10.0
